Basic shader reflection

+++ source/toaster/toast_lib/io/filesystem.hpp
@@ -16,8 +16,8 @@ namespace toaster::io::filesystem
 
 	bool exists(const Path &p_path);
 
-	std::vector<uint8> readBinary(const Path &p_path);
-	String             readFile(const Path &p_path);
+	std::vector<uint32> readBinary(const Path &p_path);
+	String              readFile(const Path &p_path);
 
 	String readFileAndSkipBOM(const Path &p_path);
 	void   writeFile(const Path &p_path, const String &p_data);
